ai-backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI integration
openai==1.3.7
tiktoken==0.5.2

# Database (MongoDB)
motor==3.3.2
pymongo==4.6.1
dnspython==2.4.2

# HTTP clients and async support
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.0

# Audio processing for TTS
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.4
scipy==1.11.4

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# WebSocket and real-time communication
websockets==12.0
sse-starlette==1.8.2

# Background tasks and caching
celery==5.3.4
redis==5.0.1

# Environment and configuration
python-dotenv==1.0.0

# Monitoring and logging
prometheus-client==0.19.0

# Database ORM (if needed for relational data)
sqlalchemy==2.0.23
alembic==1.13.1

# Additional utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
